+++ django-backend/fecfiler/reports/form_3/tests/test_serializers.py
@@ -1,13 +1,14 @@
 from django.test import TestCase
-from ..serializers import Form3XSerializer
-from fecfiler.reports.form_3.serializers import (
+from ..serializers import (
+    Form3Serializer,
     COVERAGE_DATE_REPORT_CODE_COLLISION,
     COVERAGE_DATES_EXCLUDE_EXISTING_TRANSACTIONS,
 )
+
 from fecfiler.user.models import User
 from fecfiler.reports.models import Report
 from rest_framework.request import Request, HttpRequest
-from fecfiler.reports.tests.utils import create_form3x
+from fecfiler.reports.tests.utils import create_form3
 from fecfiler.transactions.tests.utils import create_schedule_a
 from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.web_services.models import (
@@ -25,16 +26,15 @@ from fecfiler.reports.managers import (
 from fecfiler.web_services.summary.tasks import CalculationState
 from uuid import uuid4
 from datetime import datetime
-from fecfiler.web_services.summary.tasks import calculate_summary
 
 
-class F3XSerializerTestCase(TestCase):
+class F3SerializerTestCase(TestCase):
     fixtures = ["C01234567_user_and_committee"]
 
     def setUp(self):
         self.committee = CommitteeAccount.objects.create(committee_id="C00000000")
-        self.valid_f3x_report = {
-            "form_type": "F3XN",
+        self.valid_f3_report = {
+            "form_type": "F3N",
             "treasurer_last_name": "Validlastname",
             "treasurer_first_name": "Validfirstname",
             "coverage_from_date": "2023-01-01",
@@ -44,7 +44,7 @@ class F3XSerializerTestCase(TestCase):
             "upload_submission": {"fec_status": " ACCEPTED"},
         }
 
-        self.invalid_f3x_report = {
+        self.invalid_f3_report = {
             "form_type": "invalidformtype",
             "treasurer_last_name": "Validlastname",
             "date_signed": "2022-01-01",
@@ -60,13 +60,13 @@ class F3XSerializerTestCase(TestCase):
         }
 
     def test_serializer_validate(self):
-        valid_serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        valid_serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         self.assertTrue(valid_serializer.is_valid(raise_exception=True))
-        invalid_serializer = Form3XSerializer(
-            data=self.invalid_f3x_report,
+        invalid_serializer = Form3Serializer(
+            data=self.invalid_f3_report,
             context={"request": self.mock_request},
         )
         self.assertFalse(invalid_serializer.is_valid())
@@ -74,14 +74,14 @@ class F3XSerializerTestCase(TestCase):
         self.assertIsNotNone(invalid_serializer.errors["treasurer_first_name"])
 
     def test_used_report_code(self):
-        valid_serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        valid_serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         valid_serializer.is_valid()
         valid_serializer.save()
-        valid_serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        valid_serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         valid_serializer.is_valid()
@@ -90,69 +90,69 @@ class F3XSerializerTestCase(TestCase):
         )
 
     def test_get_status_mapping(self):
-        valid_serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        valid_serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
-        f3x_report = create_form3x(self.committee, "2024-01-01", "2024-02-01", {})
+        f3_report = create_form3(self.committee, "2024-01-01", "2024-02-01", {})
         # retrieve from manager to populate annotations
-        f3x_report = Report.objects.get(id=f3x_report.id)
+        f3_report = Report.objects.get(id=f3_report.id)
         valid_serializer.is_valid()
-        representation = valid_serializer.to_representation(f3x_report)
+        representation = valid_serializer.to_representation(f3_report)
         self.assertEqual(
             representation["report_status"],
             REPORT_STATUS_MAP.get(STATUS_CODE_IN_PROGRESS),
         )
 
         # .fec has been submitted but a result is pending
-        f3x_report.upload_submission = UploadSubmission.objects.initiate_submission(
-            f3x_report.id
+        f3_report.upload_submission = UploadSubmission.objects.initiate_submission(
+            f3_report.id
         )
         # retrieve from manager to populate annotations
-        f3x_report = Report.objects.get(id=f3x_report.id)
-        representation = valid_serializer.to_representation(f3x_report)
+        f3_report = Report.objects.get(id=f3_report.id)
+        representation = valid_serializer.to_representation(f3_report)
         self.assertEqual(
             representation["report_status"], REPORT_STATUS_MAP.get(STATUS_CODE_PENDING)
         )
 
         # .fec was submitted and efo came back with an 'Accepted'
-        f3x_report.upload_submission.fec_status = FECStatus.ACCEPTED
-        f3x_report.upload_submission.save()
+        f3_report.upload_submission.fec_status = FECStatus.ACCEPTED
+        f3_report.upload_submission.save()
         # retrieve from manager to populate annotations
-        f3x_report = Report.objects.get(id=f3x_report.id)
-        representation = valid_serializer.to_representation(f3x_report)
+        f3_report = Report.objects.get(id=f3_report.id)
+        representation = valid_serializer.to_representation(f3_report)
         self.assertEqual(
             representation["report_status"], REPORT_STATUS_MAP.get(STATUS_CODE_SUCCESS)
         )
 
         # an error occured at some point on our side after the user submitted
-        f3x_report.upload_submission.fecfile_task_state = FECSubmissionState.FAILED
-        f3x_report.upload_submission.fec_status = None
-        f3x_report.upload_submission.save()
+        f3_report.upload_submission.fecfile_task_state = FECSubmissionState.FAILED
+        f3_report.upload_submission.fec_status = None
+        f3_report.upload_submission.save()
         # retrieve from manager to populate annotations
-        f3x_report = Report.objects.get(id=f3x_report.id)
-        representation = valid_serializer.to_representation(f3x_report)
+        f3_report = Report.objects.get(id=f3_report.id)
+        representation = valid_serializer.to_representation(f3_report)
         self.assertEqual(
             representation["report_status"], REPORT_STATUS_MAP.get(STATUS_CODE_FAILED)
         )
 
         # .fec was submitted and efo came back with a 'rejected'
-        f3x_report.upload_submission.fecfile_task_state = FECSubmissionState.SUBMITTING
-        f3x_report.upload_submission.fec_status = FECStatus.REJECTED
-        f3x_report.upload_submission.save()
+        f3_report.upload_submission.fecfile_task_state = FECSubmissionState.SUBMITTING
+        f3_report.upload_submission.fec_status = FECStatus.REJECTED
+        f3_report.upload_submission.save()
         # retrieve from manager to populate annotations
-        f3x_report = Report.objects.get(id=f3x_report.id)
-        representation = valid_serializer.to_representation(f3x_report)
+        f3_report = Report.objects.get(id=f3_report.id)
+        representation = valid_serializer.to_representation(f3_report)
         self.assertEqual(
             representation["report_status"], REPORT_STATUS_MAP.get(STATUS_CODE_FAILED)
         )
 
     def test_update_coverage_to_overlapping_dates(self):
-        report_a = create_form3x(self.committee, "2024-01-01", "2024-03-31")
-        create_form3x(self.committee, "2024-04-01", "2024-06-30")
+        report_a = create_form3(self.committee, "2024-01-01", "2024-03-31")
+        create_form3(self.committee, "2024-04-01", "2024-06-30")
 
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         serializer.is_valid()
@@ -169,13 +169,13 @@ class F3XSerializerTestCase(TestCase):
         )
 
     def test_update_coverage_to_exclude_transaction(self):
-        report_a = create_form3x(self.committee, "2024-01-01", "2024-03-31")
+        report_a = create_form3(self.committee, "2024-01-01", "2024-03-31")
         create_schedule_a(
             "INDIVIDUAL_RECEIPT", self.committee, None, "2024-03-31", 250, report=report_a
         )
 
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         serializer.is_valid()
@@ -192,7 +192,7 @@ class F3XSerializerTestCase(TestCase):
         )
 
     def test_update_coverage_to_exclude_memo_transaction(self):
-        report_a = create_form3x(self.committee, "2024-01-01", "2024-03-31")
+        report_a = create_form3(self.committee, "2024-01-01", "2024-03-31")
         create_schedule_a(
             "INDIVIDUAL_RECEIPT",
             self.committee,
@@ -203,8 +203,8 @@ class F3XSerializerTestCase(TestCase):
             memo_code=True,
         )
 
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         serializer.is_valid()
@@ -224,7 +224,7 @@ class F3XSerializerTestCase(TestCase):
         )
 
     def test_update_coverage_to_exclude_deleted_transaction(self):
-        report_a = create_form3x(self.committee, "2024-01-01", "2024-03-31")
+        report_a = create_form3(self.committee, "2024-01-01", "2024-03-31")
         transaction = create_schedule_a(
             "INDIVIDUAL_RECEIPT",
             self.committee,
@@ -236,8 +236,8 @@ class F3XSerializerTestCase(TestCase):
         )
         transaction.delete()
 
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         serializer.is_valid()
@@ -257,13 +257,13 @@ class F3XSerializerTestCase(TestCase):
         )
 
     def test_update_coverage_dates_marks_report_dirty(self):
-        report = create_form3x(self.committee, "2026-01-01", "2026-01-31")
+        report = create_form3(self.committee, "2026-01-01", "2026-01-31")
         report.calculation_status = CalculationState.FAILED.value
         report.calculation_token = uuid4()
         report.save()
 
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
+        serializer = Form3Serializer(
+            data=self.valid_f3_report,
             context={"request": self.mock_request},
         )
         serializer.is_valid()
@@ -280,33 +280,3 @@ class F3XSerializerTestCase(TestCase):
         report.refresh_from_db()
         self.assertIsNone(report.calculation_status)
         self.assertIsNone(report.calculation_token)
-
-    def test_update_coverage_then_request_summary_updates_l6a_year_label(self):
-        report = create_form3x(self.committee, "2026-01-01", "2026-01-31")
-        calculate_summary(report.id)
-        report.refresh_from_db()
-        self.assertEqual(str(report.form_3x.L6a_year_for_above_ytd), "2026")
-
-        serializer = Form3XSerializer(
-            data=self.valid_f3x_report,
-            context={"request": self.mock_request},
-        )
-        serializer.is_valid()
-        serializer.update(
-            report,
-            {
-                "coverage_from_date": datetime.strptime("2025-01-01", "%Y-%m-%d").date(),
-                "coverage_through_date": datetime.strptime(
-                    "2025-01-31", "%Y-%m-%d"
-                ).date(),
-            },
-        )
-
-        report.refresh_from_db()
-        self.assertIsNone(report.calculation_status)
-        self.assertEqual(str(report.form_3x.L6a_year_for_above_ytd), "2026")
-
-        calculate_summary(report.id)
-        report.refresh_from_db()
-        self.assertEqual(str(report.form_3x.L6a_year_for_above_ytd), "2025")
-        self.assertEqual(report.calculation_status, CalculationState.SUCCEEDED.value)
